refactor(e2e): extract openWizardFromChoicePrompt helper, deduplicate wizard-cancel preamble in 2 test cases (-4 LOC)
